requirements-lock.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.135.3
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.4.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
limits==5.8.0
MarkupSafe==3.0.3
narwhals==2.19.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
passlib==1.7.4
pillow==12.2.0
pluggy==1.6.0
protobuf==7.34.1
pyarrow==23.0.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
smmap==5.0.3
SQLAlchemy==2.0.49
starlette==1.0.0
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
watchdog==6.0.0
wrapt==2.1.2
